Add FLUKA Hadron Inelastic Physics Constructor
This commit adds a FLUKA based hadron inelastic physics constructor.
It is imported by the work of G.Hugo for the fluka interface of
g4-11.1.ref05, but does not include HP treatment of neutrons.
When used with G4_CernFLUKAHadronInelastic_FTFP_BERT is should construct
a PL that is indetical to FTFP_BERT but uses the inelastic scattering by
FLUKA.